Fixed pre notes
