+++ feature-wallet-api/src/main/java/io/novafoundation/nova/feature_wallet_api/data/source/CoinPriceRemoteDataSource.kt
@@ -1,12 +1,13 @@
 package io.novafoundation.nova.feature_wallet_api.data.source
 
 import io.novafoundation.nova.feature_currency_api.domain.model.Currency
+import io.novafoundation.nova.feature_wallet_api.data.repository.PricePeriod
 import io.novafoundation.nova.feature_wallet_api.domain.model.CoinRateChange
 import io.novafoundation.nova.feature_wallet_api.domain.model.HistoricalCoinRate
 
 interface CoinPriceRemoteDataSource {
 
-    suspend fun getLastCoinPriceRange(priceId: String, currency: Currency, days: String): List<HistoricalCoinRate>
+    suspend fun getLastCoinPriceRange(priceId: String, currency: Currency, range: PricePeriod): List<HistoricalCoinRate>
 
     suspend fun getCoinRates(priceIds: Set<String>, currency: Currency): Map<String, CoinRateChange?>
