+++ feature-wallet-impl/src/main/java/io/novafoundation/nova/feature_wallet_impl/data/source/RealCoinPriceDataSource.kt
@@ -1,11 +1,13 @@
 package io.novafoundation.nova.feature_wallet_impl.data.source
 
 import io.novafoundation.nova.common.data.network.HttpExceptionHandler
+import io.novafoundation.nova.common.utils.KeyMutex
 import io.novafoundation.nova.common.utils.asQueryParam
 import io.novafoundation.nova.common.utils.orZero
 import io.novafoundation.nova.feature_currency_api.domain.model.Currency
 import io.novafoundation.nova.feature_wallet_api.data.network.priceApi.CoingeckoApi
 import io.novafoundation.nova.feature_wallet_api.data.network.priceApi.ProxyPriceApi
+import io.novafoundation.nova.feature_wallet_api.data.repository.PricePeriod
 import io.novafoundation.nova.feature_wallet_api.data.source.CoinPriceRemoteDataSource
 import io.novafoundation.nova.feature_wallet_api.domain.model.CoinRateChange
 import io.novafoundation.nova.feature_wallet_api.domain.model.HistoricalCoinRate
@@ -17,8 +19,15 @@ class RealCoinPriceDataSource(
     private val httpExceptionHandler: HttpExceptionHandler
 ) : CoinPriceRemoteDataSource {
 
-    override suspend fun getLastCoinPriceRange(priceId: String, currency: Currency, days: String): List<HistoricalCoinRate> {
-        val response = priceApi.getLastCoinRange(priceId, currency.coingeckoId, days)
+    private val mutex = KeyMutex()
+
+    override suspend fun getLastCoinPriceRange(priceId: String, currency: Currency, range: PricePeriod): List<HistoricalCoinRate> {
+        val key = "${priceId}_${currency.id}_$range"
+        val response = mutex.withKeyLock(key) {
+            val days = mapRangeToDays(range)
+
+            priceApi.getLastCoinRange(priceId, currency.coingeckoId, days)
+        }
 
         return response.prices.map { (timestampRaw, rateRaw) ->
             HistoricalCoinRate(
@@ -48,4 +57,12 @@ class RealCoinPriceDataSource(
     }
 
     private suspend fun <T> apiCall(block: suspend () -> T): T = httpExceptionHandler.wrap(block)
+
+    private fun mapRangeToDays(range: PricePeriod) = when (range) {
+        PricePeriod.DAY -> "1"
+        PricePeriod.WEEK -> "7"
+        PricePeriod.MONTH -> "30"
+        PricePeriod.YEAR -> "365"
+        PricePeriod.MAX -> "max"
+    }
 }
